+++ src/database/connection.js
@@ -7,9 +7,6 @@ const { Message } = require("../entities/Message");
 const { Conversation } = require("../entities/Conversation");
 const { Preferences } = require("../entities/Preferences");
 
-const { NewMessage } = require("../entities/NewMessage");
-const { NewConversation } = require("../entities/NewConversation");
-
 const connection = new Sequelize(config);
 
 User.init(connection);
@@ -17,9 +14,6 @@ Connection.init(connection);
 Preferences.init(connection);
 Message.init(connection);
 Conversation.init(connection);
-
-NewConversation.init(connection);
-NewMessage.init(connection);
 
 Connection.associate(connection.models);
 Preferences.associate(connection.models);
