+++ src/modules/messages/repositories/MessagesRepository.js
@@ -1,0 +1,5 @@
+const { Message } = require("../../../entities/Message");
+
+class MessagesRepository extends Message {}
+
+module.exports = { MessagesRepository };
