--- src/modules/messages/repositories/NewMessagesRepository.js
@@ -1,5 +1,0 @@
-const { NewMessage } = require("../../../entities/NewMessage");
-
-class NewMessagesRepository extends NewMessage {}
-
-module.exports = { NewMessagesRepository };
