--- src/modules/users/repositories/NewConversationRepository.js
@@ -1,5 +1,0 @@
-const { NewConversation } = require("../../../entities/NewConversation");
-
-class NewConversationRepository extends NewConversation {}
-
-module.exports = { NewConversationRepository };
