--- src/modules/users/services/ListAllConversationsUserService.js
@@ -1,16 +1,0 @@
-const {
-  ConversationsRepository,
-} = require("../repositories/ConversationsRepository");
-
-class ListAllConversationsUserService {
-  async handle(fkUserSender) {
-    const conversations = await ConversationsRepository.findAll({
-      where: { fkUserSender },
-      include: [{ association: "user_receiver" }],
-    });
-
-    return conversations;
-  }
-}
-
-module.exports = { ListAllConversationsUserService };
