+++ src/websockets/user.js
@@ -3,9 +3,9 @@ const logoutUserUseCase = require("../modules/users/useCases/LogoutUserUseCase")
 const listAllUsersConnectionUseCase = require("../modules/connections/useCases/ListAllUsersConnectionUseCase");
 
 // Novos useCases para conversas e mensagens (Nova feature)
-const sendNewMessageUseCase = require("../modules/users/useCases/SendNewMessageUseCase");
-const listLastNewConversations = require("../modules/users/useCases/ListLastNewConversationUseCase");
-const listNewMessagesUseCase = require("../modules/messages/useCases/ListAllNewMessagesUseCase");
+const sendMessageUseCase = require("../modules/users/useCases/SendMessageUseCase");
+const listLastConversationsUseCase = require("../modules/users/useCases/ListLastConversationUseCase");
+const listAllMessagesUseCase = require("../modules/messages/useCases/ListAllMessagesUseCase");
 
 module.exports = () => {
   global.io.on("connect", (socket) => {
@@ -22,10 +22,10 @@ module.exports = () => {
     });
 
     // Novos eventos para conversas e mensagens (Nova feature)
-    socket.on("user_send_new_message", sendNewMessageUseCase);
+    socket.on("user_send_new_message", sendMessageUseCase);
 
-    socket.on("list_user_new_conversations", listLastNewConversations);
+    socket.on("list_user_new_conversations", listLastConversationsUseCase);
 
-    socket.on("list_new_messages", listNewMessagesUseCase);
+    socket.on("list_new_messages", listAllMessagesUseCase);
   });
 };
